requirements.txt
absl-py==1.4.0
aiohttp==3.8.1
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
blinker==1.5
brotlipy==0.7.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.14.6
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.5
cryptography==37.0.2
cycler==0.11.0
dropblock==0.3.0
einops==0.6.1
fonttools==4.25.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.1
google-auth-oauthlib==0.4.6
grpcio==1.42.0
idna==3.3
importlib-metadata==6.0.0
importlib-resources==5.12.0
kiwisolver==1.4.4
Markdown==3.4.1
matplotlib==3.7.0
multidict==6.0.2
munkres==1.1.4
numpy==1.22.3
oauthlib==3.2.2
olefile==0.46
opencv-python==4.7.0.72
packaging==23.0
pandas==1.4.2
patsy==0.5.3
Pillow==8.2.0
pip==22.1.2
ply==3.11
protobuf==3.15.8
psutil==5.9.0
ptflops==0.7
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-model-summary==0.1.2
pytz==2022.7.1
pyu2f==0.1.5
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.5.3
seaborn==0.12.2
setuptools==67.4.0
sip==6.6.2
six==1.16.0
smmap==3.0.5
statsmodels==0.13.5
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop
toml==0.10.2
tornado==6.1
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.1.2
wheel==0.37.1
yarl==1.7.2
zipp==3.15.0
